DRTII-1123 Move serialiser test here
